# assets/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# ChatPuppy Api Document\n\n## Auth\n- Add headers `Authorization Bearer {token}`\n- Add headers `X-Member-Address ${address}`\n\n## Signature\n1. Just post request\n2. Encrypt raw post body string to `MD5`, getting string `bodyMD5`\n3. Concatenate `timestamp` (from 1970, 10 digits) string end of step 2, getting string `s = bodyMD5 + timestamp`\n4. Sign `s` getting `signature` string\n5. Add headers `X-Signature ${signature}`\n6. Add headers `X-Timestamp ${timestamp}`\n\n## WebSocket\n\u003e `/message`\n\n### Operates\n- 1: register\n- 2: set group key\n- 3: new group message\n",
        "title": "ChatPuppy Api Doc",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/group": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Filter group",
                "operationId": "GroupList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group's category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "current page, start from 1 (default: 1)",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count peer page, default is 50",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/model.PaginationRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.GroupListRes"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "create an group (need signature verify)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Create group",
                "operationId": "GroupCreate",
                "parameters": [
                    {
                        "description": "Group info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GroupDetailWithPublicKey"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/active": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Active group",
                "operationId": "GroupActive",
                "parameters": [
                    {
                        "description": "Group id",
                        "name": "groupId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/group/join": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Join group",
                "operationId": "GroupJoin",
                "parameters": [
                    {
                        "description": "Join info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupJoinReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GroupDetailWithPublicKey"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/group/key": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Share key with group",
                "operationId": "GroupShareKey",
                "parameters": [
                    {
                        "description": "Group and key info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupShareKeyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GroupPublicKey"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/group/key/used": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Setting current key of group",
                "operationId": "GroupKeyUsed",
                "parameters": [
                    {
                        "description": "Group keys",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupKeyUsedReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/group/kickout": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Kickout member",
                "operationId": "GroupKickout",
                "parameters": [
                    {
                        "description": "Group member info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/group/leave": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Leave group",
                "operationId": "GroupLeave",
                "parameters": [
                    {
                        "description": "Group id",
                        "name": "groupId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/group/manager": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Set manager",
                "operationId": "GroupManager",
                "parameters": [
                    {
                        "description": "Group member info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/group/recode": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Regenerate invite code",
                "operationId": "GroupReCode",
                "parameters": [
                    {
                        "description": "Group id",
                        "name": "groupId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/group/update": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Update group info",
                "operationId": "GroupUpdate",
                "parameters": [
                    {
                        "description": "Group info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GroupUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/group/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Detail info of group",
                "operationId": "GroupDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GroupDetailRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.GroupCreateReq": {
            "type": "object",
            "required": [
                "category",
                "name",
                "sharedPublic"
            ],
            "properties": {
                "category": {
                    "description": "group's category",
                    "type": "string"
                },
                "intro": {
                    "description": "group intro",
                    "type": "string"
                },
                "maxMembers": {
                    "description": "group's max members",
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2
                },
                "public": {
                    "description": "`true` create public group, `false` create private group",
                    "type": "boolean"
                },
                "sharedPublic": {
                    "description": "Member's public key for `ecdh` exchange",
                    "type": "string"
                }
            }
        },
        "model.GroupDetailRes": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "intro": {
                    "description": "optional",
                    "type": "string"
                },
                "inviteCode": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MemberWithPermission"
                    }
                },
                "membersCount": {
                    "type": "integer"
                },
                "membersMax": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "description": "group is belongs to current member",
                    "type": "boolean"
                },
                "public": {
                    "description": "group is public or private",
                    "type": "boolean"
                }
            }
        },
        "model.GroupDetailWithPublicKey": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "groupPublicKey": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "intro": {
                    "description": "optional",
                    "type": "string"
                },
                "inviteCode": {
                    "type": "string"
                },
                "keyId": {
                    "type": "string"
                },
                "membersCount": {
                    "type": "integer"
                },
                "membersMax": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "description": "group is belongs to current member",
                    "type": "boolean"
                },
                "public": {
                    "description": "group is public or private",
                    "type": "boolean"
                }
            }
        },
        "model.GroupJoinReq": {
            "type": "object",
            "required": [
                "sharedPublic"
            ],
            "properties": {
                "groupId": {
                    "type": "string"
                },
                "inviteCode": {
                    "description": "invite code from group member",
                    "type": "string"
                },
                "sharedPublic": {
                    "description": "Member's public key for `ecdh` exchange",
                    "type": "string"
                }
            }
        },
...
        "model.GroupKeyUsed": {
            "type": "object",
            "required": [
                "groupId",
                "keyId"
            ],
            "properties": {
                "groupId": {
                    "type": "string"
                },
                "keyId": {
                    "type": "string"
                }
            }
        },
        "model.GroupKeyUsedReq": {
            "type": "object",
            "required": [
                "keys"
            ],
            "properties": {
                "keys": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.GroupKeyUsed"
                    }
                }
            }
        },
        "model.GroupListRes": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "intro": {
                    "description": "optional",
                    "type": "string"
                },
                "joined": {
                    "type": "boolean"
                },
                "membersCount": {
                    "type": "integer"
                },
                "membersMax": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.GroupMemberReq": {
            "type": "object",
            "properties": {
                "groupId": {
                    "type": "string"
                },
                "memberId": {
                    "type": "string"
                }
            }
        },
        "model.GroupPublicKey": {
            "type": "object",
            "properties": {
                "groupPublicKey": {
                    "type": "string"
                },
                "keyId": {
                    "type": "string"
                }
            }
        },
        "model.GroupShareKeyReq": {
            "type": "object",
            "required": [
                "groupId",
                "sharedPublic"
            ],
            "properties": {
                "groupId": {
                    "description": "group id",
                    "type": "string"
                },
                "sharedPublic": {
                    "description": "Member's public key for `ecdh` exchange",
                    "type": "string"
                }
            }
        },
        "model.GroupUpdateReq": {
            "type": "object",
            "required": [
                "groupId"
            ],
            "properties": {
                "category": {
                    "description": "group's category",
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "intro": {
                    "type": "string"
                },
                "maxMembers": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                }
            }
        },
...
        "model.MemberWithPermission": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "permission": {
                    "type": "integer"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.PaginationRes": {
            "type": "object",
            "properties": {
                "items": {},
                "pagination": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Pagination"
                        }
                    ]
                }
            }
        }
    }
}